+++ tests/Field/Helper/FieldDefinitionHelperTest.php
@@ -3,6 +3,7 @@
 namespace Tests\Torr\Storyblok\Field\Helper;
 
 use Torr\Storyblok\Exception\InvalidComponentConfigurationException;
+use Torr\Storyblok\Field\Definition\AssetField;
 use Torr\Storyblok\Field\Definition\TextField;
 use Torr\Storyblok\Field\Helper\FieldDefinitionHelper;
 use PHPUnit\Framework\TestCase;
@@ -39,7 +40,7 @@ class FieldDefinitionHelperTest extends TestCase
 	/**
 	 *
 	 */
-	public function testMultiple () : void
+	public function testMultipleNames () : void
 	{
 		$this->expectException(InvalidComponentConfigurationException::class);
 		$this->expectExceptionMessage("Can't use multiple fields as admin display name");
@@ -49,6 +50,22 @@ class FieldDefinitionHelperTest extends TestCase
 				->useAsAdminDisplayName(),
 			"two" => (new TextField("Text 2"))
 				->useAsAdminDisplayName(),
+		]);
+	}
+
+	/**
+	 *
+	 */
+	public function testMultipleImages () : void
+	{
+		$this->expectException(InvalidComponentConfigurationException::class);
+		$this->expectExceptionMessage("Can't use multiple fields as admin display image");
+
+		FieldDefinitionHelper::ensureMaximumOneAdminDisplayName([
+			"one" => (new AssetField("Image 1"))
+				->useAsAdminDisplayImage(),
+			"two" => (new AssetField("Image 2"))
+				->useAsAdminDisplayImage(),
 		]);
 	}
 }